Data: Move registry into own file
